+++ app/editor/page.tsx
@@ -1180,20 +1180,29 @@ export default function Home() {
                                     "\n"
                                   );
                                   const heads = match(lines.shift() || "");
-                                  const data: Record<string, string>[] =
-                                    lines.map((line) => {
-                                      return match(line).reduce(
-                                        (
-                                          acc: Record<string, string>,
-                                          cur,
-                                          i
-                                        ) => {
-                                          const key = heads[i] ?? `extra_${i}`;
-                                          return { ...acc, [key]: cur };
-                                        },
-                                        {}
-                                      );
-                                    });
+                                  const data: Record<
+                                    string,
+                                    string | number | null
+                                  >[] = lines.map((line) => {
+                                    return match(line).reduce(
+                                      (
+                                        acc: Record<
+                                          string,
+                                          string | number | null
+                                        >,
+                                        cur,
+                                        i
+                                      ) => {
+                                        const val =
+                                          cur.length <= 0
+                                            ? null
+                                            : Number(cur) || cur;
+                                        const key = heads[i] ?? `extra_${i}`;
+                                        return { ...acc, [key]: val };
+                                      },
+                                      {}
+                                    );
+                                  });
                                   if (overpage.type == "importing") {
                                     heads
                                       .filter(
@@ -1229,7 +1238,7 @@ export default function Home() {
                                             option: string;
                                             categories: Record<string, number>;
                                           } = {
-                                            option: val.Text,
+                                            option: String(val.Text),
                                             categories: {},
                                           };
                                           delete val.Text;
@@ -1237,7 +1246,7 @@ export default function Home() {
                                           option.categories = Object.entries(
                                             val
                                           )
-                                            .filter(([, val]) => Number(val))
+                                            .filter(([, val]) => val)
                                             .reduce(
                                               (
                                                 res: Record<string, number>,
@@ -1266,7 +1275,7 @@ export default function Home() {
                                             type: translator[
                                               val["Question Type"]
                                             ],
-                                            name: val.Text,
+                                            name: String(val.Text),
                                             values: [],
                                           });
                                         }
